refactor satisfaction DNA in planning zome into modular crates
minor code changes: vf_planning -> vf_core
