+++ Cargo.toml
@@ -37,6 +37,11 @@ members= [
   "lib/rea_resource_specification/storage",
   "lib/rea_resource_specification/storage_consts",
   "lib/rea_resource_specification/rpc",
+  "lib/rea_satisfaction/defs",
+  "lib/rea_satisfaction/lib",
+  "lib/rea_satisfaction/storage",
+  "lib/rea_satisfaction/storage_consts",
+  "lib/rea_satisfaction/rpc",
   "lib/rea_unit/defs",
   "lib/rea_unit/lib",
   "lib/rea_unit/storage",
