+++ happs/planning/zomes/satisfaction/code/.hcbuild
@@ -10,5 +10,5 @@
       ]
     }
   ],
-  "artifact": "/tmp/holochain/target/wasm32-unknown-unknown/release/rea_planning_satisfaction.wasm"
+  "artifact": "/tmp/holochain/target/wasm32-unknown-unknown/release/hc_zome_rea_satisfaction.wasm"
 }
