+++ lib/rea_commitment/defs/Cargo.toml
@@ -17,6 +17,7 @@ hc_zome_rea_commitment_storage = { path = "../storage" }
 hc_zome_rea_commitment_storage_consts = { path = "../storage_consts" }
 hc_zome_rea_commitment_rpc = { path = "../rpc" }
 hc_zome_rea_process_storage_consts = { path = "../../rea_process/storage_consts" }
+hc_zome_rea_satisfaction_storage_consts = { path = "../../rea_satisfaction/storage_consts" }
 
 [lib]
 crate-type = ["lib"]
