+++ lib/rea_economic_event/defs/src/lib.rs
@@ -11,10 +11,9 @@ use hdk::prelude::*;
 
 use vf_planning::identifiers::{
     FULFILLMENT_BASE_ENTRY_TYPE,
-    SATISFACTION_BASE_ENTRY_TYPE,
 };
-
 use hc_zome_rea_process_storage_consts::PROCESS_BASE_ENTRY_TYPE;
+use hc_zome_rea_satisfaction_storage_consts::SATISFACTION_BASE_ENTRY_TYPE;
 use hc_zome_rea_economic_event_storage_consts::*;
 use hc_zome_rea_economic_event_storage::{
     Entry as EconomicEventEntry,
