+++ lib/vf_planning/src/lib.rs
@@ -13,7 +13,6 @@ extern crate vf_core;
 pub use vf_core::{ type_aliases, measurement };
 
 pub mod fulfillment;
-pub mod satisfaction;
 
 pub mod identifiers {
     pub const BRIDGED_OBSERVATION_DHT: &str = "vf_observation";
@@ -26,11 +25,4 @@ pub mod identifiers {
     pub const FULFILLMENT_FULFILLEDBY_LINK_TYPE: &str = "vf_fulfillment_fulfilled_by";
     pub const FULFILLMENT_FULFILLEDBY_LINK_TAG: &str = "fulfilled_by";
 
-    pub const SATISFACTION_BASE_ENTRY_TYPE: &str = "vf_satisfaction_baseurl";
-    pub const SATISFACTION_INITIAL_ENTRY_LINK_TYPE: &str = "vf_satisfaction_entry";
-    pub const SATISFACTION_ENTRY_TYPE: &str = "vf_satisfaction";
-    pub const SATISFACTION_SATISFIES_LINK_TYPE: &str = "vf_satisfaction_satisfies";
-    pub const SATISFACTION_SATISFIES_LINK_TAG: &str = "satisfies";
-    pub const SATISFACTION_SATISFIEDBY_LINK_TYPE: &str = "vf_satisfaction_satisfied_by";
-    pub const SATISFACTION_SATISFIEDBY_LINK_TAG: &str = "satisfied_by";
 }
